test(jsdom): update jest
