+++ .dependency-cruiser.json
@@ -77,7 +77,7 @@
       "comment": "Don't allow dependencies from src/app/lib to a development only package",
       "from": {
         "path": "^(src|app|lib)",
-        "pathNot": "\\.spec\\.(js|ts|ls|coffee|litcoffee|coffee\\.md)$"
+        "pathNot": "__tests__"
       },
       "to": {
         "dependencyTypes": [
