+++ package.json
@@ -73,6 +73,7 @@
     "@types/fs-extra": "^5.0.5",
     "@types/html-webpack-plugin": "3.2.0",
     "@types/jest": "^24.0.6",
+    "@types/jsdom": "^12.2.3",
     "@types/mini-css-extract-plugin": "^0.2.0",
     "@types/node": "^11.9.5",
     "@types/optimize-css-assets-webpack-plugin": "^1.3.4",
@@ -98,6 +99,8 @@
     "husky": "^1.3.1",
     "jest": "^24.1.0",
     "jest-cli": "^24.1.0",
+    "jest-pnp-resolver": "^1.2.0",
+    "jsdom": "^13.2.0",
     "lint-staged": "^8.1.4",
     "mini-css-extract-plugin": "^0.5.0",
     "optimize-css-assets-webpack-plugin": "^5.0.1",
